backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coppermind"
version = "1.0.0"
description = "Copper market sentiment analysis and price prediction system"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "CopperMind Team"}
]

dependencies = [
    # Web framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    
    # Data fetching
    "yfinance>=0.2.33",
    "requests>=2.31.0",
    "feedparser>=6.0.10",
    
    # NLP & ML
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "xgboost>=2.0.2",
    "scikit-learn>=1.3.2",
    
    # Data processing
    "pandas>=2.1.3",
    "numpy>=1.26.2",
    
    # Text processing
    "langdetect>=1.0.9",
    "rapidfuzz>=3.5.2",
    "beautifulsoup4>=4.12.2",
    
    # Scheduling
    "apscheduler>=3.10.4",
    
    # Utilities
    "python-dateutil>=2.8.2",
    "filelock>=3.13.1",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

